+++ app/src/main/java/com/example/senla_tz/util/extends/LatLngExtends.kt
@@ -3,7 +3,7 @@ package com.example.senla_tz.util.extends
 import android.location.Location
 import com.google.android.gms.maps.model.LatLng
 
-fun List<LatLng>.distance(): Int{
+fun List<LatLng>.distance(): Long{
     var distance = 0f
 
     reduce{ firstLatLng,  secondLatLng ->
@@ -17,5 +17,5 @@ fun List<LatLng>.distance(): Int{
         secondLatLng
     }
 
-    return distance.toInt()
+    return distance.toLong()
 }
